+++ apps/web/src/components/editor/dialogs/FormulaEditorDialog.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import { escapeHtml, normalizeFormulaInput, wrapFormula } from '@/lib/markdown/formula'
+import { escapeHtml } from '@md/core'
+import { normalizeFormulaInput, wrapFormula } from '@/lib/markdown/formula'
 import { isMathJaxLoaded, loadMathJax } from '@/lib/preview/mathjax'
 import { useEditorStore } from '@/stores/editor'
 import { useUIStore } from '@/stores/ui'
