+++ apps/web/src/composables/useInitialPreviewBoot.ts
@@ -1,4 +1,5 @@
 import { dismissInitialLoader } from '@/lib/bootstrap/dismiss-initial-loader'
+import { delay } from '@/lib/delay'
 import { waitForPreviewReady } from '@/lib/preview/preview-ready'
 
 const INITIAL_PREVIEW_TIMEOUT_MS = 20_000
@@ -6,10 +7,6 @@ const OUTPUT_WAIT_TIMEOUT_MS = 5_000
 const OUTPUT_POLL_INTERVAL_MS = 50
 
 let initialPreviewBootstrapped = false
-
-function delay(ms: number) {
-  return new Promise<void>(resolve => window.setTimeout(resolve, ms))
-}
 
 /** Wait for PreviewPanel mount and #output in the DOM. */
 async function waitForOutputElement(timeoutMs = OUTPUT_WAIT_TIMEOUT_MS) {
