+++ apps/web/src/lib/delay.ts
@@ -1,0 +1,4 @@
+/** Promise-based delay backed by window.setTimeout. */
+export function delay(ms: number) {
+  return new Promise<void>(resolve => window.setTimeout(resolve, ms))
+}
